requirements/base.txt
# Core Framework Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database & ORM
SQLAlchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Cache & Message Queue
redis>=5.0.1
redis[hiredis]>=5.0.1
aio-pika>=9.3.1

# HTTP Client
aiohttp>=3.9.1
httpx>=0.25.2

# Security & Encryption
cryptography>=41.0.8
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
werkzeug>=3.0.1

# Environment Configuration
python-dotenv>=1.0.0

# Observability & Monitoring
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-instrumentation-requests>=0.42b0
opentelemetry-instrumentation-aiohttp-client>=0.42b0
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0

# Payment Providers
stripe>=7.8.0
razorpay>=1.4.2

# Validation & Serialization
marshmallow>=3.20.2
phonenumbers>=8.13.26

# Utilities
python-multipart>=0.0.6
email-validator>=2.1.0
Pillow>=10.1.0
psutil>=5.9.6

# Background Tasks
celery>=5.3.4
flower>=2.0.1

# Data Science & Analytics
pandas>=2.1.4
numpy>=1.26.2
plotly>=5.17.0
dash>=2.14.2

# Console and Logging
rich>=13.7.0
typer>=0.9.0
click>=8.1.7

# Utilities & CLI
tenacity>=8.2.3
